Add files via upload
In recent years, feature selection research has quickly advanced to keep up with the age of developing expert systems. This is because the applications of these systems sometimes need massive datasets. Researchers who have an interest in creating novel feature selection methods or enhancing existing technologies have grown their interest in this topic. The current version of binary PSO (BPSO) is not developed as well as continuous PSO and does not follow the main principles of the standard PSO algorithm. Unlike the continuous version of the PSO, in BPSO, particle position is restricted to the Hamming space. Thus, there is no risk of swarm divergence, but the problem of premature convergence of the swarm arises. This paper presents an enhanced particle swarm optimization employed for feature selection to tackle the limitations of the current version of binary BPSO. Its primary quality is the application of a position update mechanism to appropriately select optimal features. The position update mechanism is designed to avoid the premature convergence issue suffered by BPSO, by determining the probability of the particles switching positions using position values rather than velocity. Two variant methods for determining the probability of changing the position of a particle element were introduced. These result in the two variants of enhanced binary particle swarm optimization (EBPSO) for feature selection, called EBPSO1 and EBPSO2. Several experiments were performed to analyze the effectiveness of the proposed method and compare it with state-of-the-art feature selection methods based on eighteen frequently used benchmark datasets.